Fixed some Warnings about non-initialized variables.

+++ include/fcl/BV/kIOS.h
@@ -68,7 +68,7 @@ class kIOS
     }
     else /** spheres partially overlapping or disjoint */
     {
-      float dist = std::sqrt(dist2);
+      float dist = static_cast<float>(std::sqrt(dist2) );
       kIOS_Sphere s;
       s.r = dist + s0.r + s1.r;
       if(dist > 0)
